extensions/reaper/requirements.txt
shiny
python-dotenv
posit-sdk
polars
